+++ src/main/java/org/drugis/addis/trialverse/service/TrialverseService.java
@@ -8,5 +8,5 @@ import java.util.List;
  * Created by connor on 25-3-14.
  */
 public interface TrialverseService {
-  public List<JSONObject> getVariablesByOutcomeIds(List<Integer> outcomeIds);
+  public List<JSONObject> getVariablesByOutcomeIds(List<Long> outcomeIds);
 }
